pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-mcp-server"
version = "0.1.0"
description = "MCP server for reading and sending emails via IMAP/SMTP"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "aiostream>=0.4.5",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
]
